+++ pytorch_lightning/accelerators/ddp_base_backend.py
@@ -19,7 +19,8 @@ import torch.distributed as torch_distrib
 import torch.distributed as dist
 
 from pytorch_lightning import _logger as log
-from pytorch_lightning.accelerators.base_backend import Accelerator, BackendType, DeviceType
+from pytorch_lightning.accelerators.base_backend import Accelerator
+from pytorch_lightning.accelerators.types import BackendType, DeviceType
 from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.utilities.cloud_io import atomic_save
 from pytorch_lightning.utilities.distributed import rank_zero_only, rank_zero_warn
