+++ tests/base/deterministic_model.py
@@ -3,7 +3,7 @@ from torch import nn
 from torch.utils.data import Dataset, DataLoader
 
 from pytorch_lightning import TrainResult, EvalResult
-from pytorch_lightning.accelerators.base_backend import BackendType
+from pytorch_lightning.accelerators.types import BackendType
 from pytorch_lightning.core.lightning import LightningModule
