+++ tests/models/test_tpu.py
@@ -5,7 +5,7 @@ from torch.utils.data import DataLoader
 
 import tests.base.develop_pipelines as tpipes
 from pytorch_lightning import Trainer, seed_everything
-from pytorch_lightning.accelerators.base_backend import BackendType, DeviceType
+from pytorch_lightning.accelerators.types import BackendType, DeviceType
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
 from tests.base.datasets import TrialMNIST
